migrate modules_test_status to pytest
